Merge v2.0.0-beta.1 release lineage into main

+++ app/apps/app_waveform/app_waveform.cpp
@@ -7,7 +7,7 @@
 #include "../../hal/hal.h"
 #include "../utils/system/system.h"
 #include "../../assets/assets.h"
-#include "../app_files/app_files.h"
+#include "../app_settings/app_settings.h"
 #include "apps/utils/system/inputs/button/button.h"
 #include "spdlog/spdlog.h"
 #include <string>
@@ -17,7 +17,46 @@ using namespace MOONCAKE::APPS;
 using namespace SYSTEM::INPUTS;
 using namespace SYSTEM::UI;
 
+namespace
+{
+    WEB_SERVER_OWNER::StartResult StartSystemLive(void* context)
+    {
+        if (context == nullptr)
+            return WEB_SERVER_OWNER::StartResult::RouteOrRegistrationFailure;
+        return HAL::StartSystemLiveSharing(
+            *static_cast<const VIEWER_ASSET_CONTRACT::DisplayProfile*>(context));
+    }
+
+    LIVE_SHARE_SESSION::TransportStopStatus BeginSystemLiveStop(void*) { return HAL::BeginSystemLiveStop(); }
+
+    LIVE_SHARE_SESSION::TransportStopStatus PollSystemLiveStop(void*) { return HAL::PollSystemLiveStop(); }
+
+    WEB_SERVER_OWNER::StopResult FinishSystemLiveStop(void*) { return HAL::FinishSystemLiveStop(); }
+
+    std::uint32_t CurrentMillis(void*) { return HAL::Millis(); }
+
+    std::uint8_t CurrentStationCount(void*) { return HAL::GetApStaNum(); }
+
+    LIVE_SHARE_CONTROLLER::TransportCallbacks
+    MakeLiveShareCallbacks(VIEWER_ASSET_CONTRACT::DisplayProfile* displayProfile)
+    {
+        LIVE_SHARE_CONTROLLER::TransportCallbacks callbacks;
+        callbacks.context = displayProfile;
+        callbacks.startSystemLive = StartSystemLive;
+        callbacks.beginSystemLiveStop = BeginSystemLiveStop;
+        callbacks.pollSystemLiveStop = PollSystemLiveStop;
+        callbacks.finishSystemLiveStop = FinishSystemLiveStop;
+        callbacks.millis = CurrentMillis;
+        callbacks.stationCount = CurrentStationCount;
+        return callbacks;
+    }
+} // namespace
+
 AppWaveform::WaveformMode_t AppWaveform::_mode = AppWaveform::mode_both;
+
+static_assert(AppWaveform::mode_both == 0, "mode_both display-profile contract");
+static_assert(AppWaveform::mode_volt_only == 1, "mode_volt_only display-profile contract");
+static_assert(AppWaveform::mode_current_only == 2, "mode_current_only display-profile contract");
 
 const char* AppWaveform_Packer::getAppName() { return AssetPool::GetText().AppName_Waveform; }
 
@@ -33,6 +72,10 @@ void AppWaveform::onResume()
     spdlog::info("{} onResume", getAppName());
     _data.view = new VIEWS::WaveFormRecorder(AssetPool::GetColor().AppWaveform.primary, (int)_mode);
     _data.view->init();
+    _data.display_profile =
+        VIEWER_ASSET_CONTRACT::DisplayProfileForWaveformModeCode(static_cast<std::uint8_t>(_mode));
+    _data.live_share_controller =
+        new LIVE_SHARE_CONTROLLER::LiveShareController(MakeLiveShareCallbacks(&_data.display_profile));
 
     // Footprint
     HAL::NvsSet(NVS_KEY_APP_HISTORY, 5);
@@ -41,7 +84,39 @@ void AppWaveform::onResume()
 // Like loop()...
 void AppWaveform::onRunning()
 {
-    _data.view->update();
+    Button::Update();
+    Encoder::Update();
+
+    LIVE_SHARE_CONTROLLER::InputSnapshot input;
+    input.sideClicked = Button::Side()->wasClicked();
+    input.sideHeld = Button::Side()->wasHold();
+    input.encoderClicked = Button::Encoder()->wasClicked();
+    input.encoderHeld = Button::Encoder()->wasHold();
+
+    if (_data.live_share_controller->ownsInteraction())
+    {
+        _data.live_share_controller->update(input);
+        _sync_live_share_view();
+        if (_data.live_share_controller->ownsInteraction())
+            _render_live_share_view();
+        return;
+    }
+
+    const LIVE_SHARE_CONTROLLER::RecorderActivity recorderActivity = _data.view->activity();
+    const LIVE_SHARE_CONTROLLER::ForegroundAction action = LIVE_SHARE_CONTROLLER::ResolveWaveformInput(
+        input, recorderActivity, _data.live_share_controller->state(), _data.view->recordingInputAvailable());
+
+    _data.view->updateForeground(input, action, _data.live_share_controller->isInactive());
+
+    if (action == LIVE_SHARE_CONTROLLER::ForegroundAction::StartLiveView)
+    {
+        _data.live_share_controller->requestStart(_data.view->activity());
+        spdlog::info("Live View start result: {}",
+                     LIVE_SHARE_SESSION::StartOutcomeName(_data.live_share_controller->lastStartOutcome()));
+        _sync_live_share_view();
+        if (_data.live_share_controller->ownsInteraction())
+            _render_live_share_view();
+    }
 
     // Check kill signal
     if (_data.view->want2quit())
@@ -54,9 +129,36 @@ void AppWaveform::onRunning()
 
 void AppWaveform::onDestroy()
 {
+    (void)HAL::TerminateMeasurementSession(LIVE_SHARE_SAFETY::TerminationReason::MeasurementExit);
+    if (_data.live_share_controller)
+        _data.live_share_controller->measurementTerminationObserved(LIVE_SHARE_SESSION::MeasurementTermination::Exit);
     spdlog::info("{} onDestroy", getAppName());
     delete _data.view;
+    delete _data.live_share_view;
+    delete _data.live_share_controller;
     NotificationBubble::Free();
+}
+
+void AppWaveform::_sync_live_share_view()
+{
+    if (_data.live_share_controller->ownsInteraction() && _data.live_share_view == nullptr)
+        _data.live_share_view = new VIEWS::LiveShareView;
+    else if (_data.live_share_controller->isInactive() && _data.live_share_view != nullptr)
+    {
+        delete _data.live_share_view;
+        _data.live_share_view = nullptr;
+    }
+}
+
+void AppWaveform::_render_live_share_view()
+{
+    if (_data.live_share_view == nullptr)
+        return;
+    _data.live_share_view->update(_data.live_share_controller->state(),
+                                  AssetPool::GetColor().AppWaveform.primary,
+                                  HAL::GetSystemLiveWifiSsid(),
+                                  HAL::GetSystemLiveViewerUrl(),
+                                  _data.live_share_controller->lastStartOutcome());
 }
 
 void AppWaveform::_handle_recording_finished()
@@ -65,10 +167,10 @@ void AppWaveform::_handle_recording_finished()
     delete _data.view;
 
     // What to do with
-    // std::vector<std::string> option_list = {" - Open", " - Upload (EzData)", " - Delete", " - Continue"};
+    // std::vector<std::string> option_list = {" - Open", " - Network", " - Delete", " - Continue"};
     std::vector<std::string> option_list;
     option_list.push_back(AssetPool::GetText().AppFiles_Option_Open);
-    option_list.push_back(AssetPool::GetText().AppFiles_Option_Upload);
+    option_list.push_back(AssetPool::GetText().AppSettings_Option_Network);
     option_list.push_back(AssetPool::GetText().AppFiles_Option_Delete);
     option_list.push_back(AssetPool::GetText().AppSettings_Option_Back);
 
@@ -94,17 +196,20 @@ void AppWaveform::_handle_recording_finished()
             VaRecordViewer::CreateAndWait(&record);
         }
 
-        // Download (QR)
+        // Network
         else if (selected_index == 1)
         {
-            _on_page_download_local(HAL::GetLatestVaRecordName());
+            AppSettings::_on_page_network();
         }
 
         // Delete
         else if (selected_index == 2)
         {
-            HAL::DeleteVaRecord(HAL::GetLatestVaRecordName());
-            break;
+            if (CreateConfirmPage(AssetPool::GetText().AppFiles_Confirm_Delete, false))
+            {
+                HAL::DeleteVaRecord(HAL::GetLatestVaRecordName());
+                break;
+            }
         }
     }
 
@@ -112,5 +217,3 @@ void AppWaveform::_handle_recording_finished()
     _data.view = new VIEWS::WaveFormRecorder(AssetPool::GetColor().AppWaveform.primary, (int)_mode);
     _data.view->init();
 }
-
-void AppWaveform::_on_page_download_local(const std::string& recordName) { SYSTEM::UI::CreateDownloadQRPage(recordName); }
